+++ package.json
@@ -15,6 +15,7 @@
     "@actions/github": "^4.0.0",
     "@octokit/rest": "^18.0.0",
     "clone": "^2.1.2",
+    "https-proxy-agent": "^5.0.0",
     "js-yaml": "^3.14.0",
     "mustache": "^4.0.1",
     "sanitize-filename": "^1.6.3",
